Resolve IllegalArgumentException being thrown when setting current activity
S2-group/NAPPA#53

This code relies on having the activity ID.
The activity ID is only created after inserting the data in the database.
The data is only inserted in the database afterward.
As such, in the original implementation, the observers of a new activity was never added,
since the ID was unknown at that moment.