Fix invoking trigger to register the same activity multiple times.
Currently, the register activity handler was triggered whenever an `ActivityNode` object was instantiated. This caused the register activity handler to be triggered on multiple occasions -- even when registering the activity. The trigger invocation code block was moved to the navigation monitor, thus invoking it only once per navigation and executing only if the activity is not registered.