Allow apps to pass Intent Extras with a null value
One of the apps selected as subjects for the experiment was passing null values using Intent Extras. There is no rule in Android forbidding to doing so. Therefore, there is no reason for NAPPA to throw an Exception when parsing an Intent Extra with a null value. Instead, NAPPA should just ignore this extra and continue the execution.