Improvements and fixes to Greedy strategy in time and frequency
1) Only run score calculations if needed.
If the current node doesn't have any successor or it has only a single successor, then there is no need to run the score calculations. For the first case (no successor), end the recursion and return the URL list selected so far. For the second case (single successor), this node will be the best successor and its score will be equal to the parent score. The successor score will receive 90% of the parent score in order to ensure that we don't traverse too deeply in the graph when running the strategy for a list of nodes (i.e., the path in the graph is composed of nodes with a single child). In this scenario, by multiplying the scores by 90% on each recursion, there will be at most 5 consecutive recursions as 0.9^5 < 0.6. The value of 90% was arbitrarily selected.

2) If a node was already visited, terminate the recursion
If the current path in the graph is part of a (small) closed-loop, then the recursion will visit the same nodes multiple times. The previous item partially solves this issue if all nodes in the loop have a single child only. However, this is not always the case. Therefore, to avoid repeating calculations, URLs and recursions, keep a list of visited nodes. If the current node in the recursion has already been visited, end the recursion and return the URL list selected so far.

3) Reduce recursion stack.
There is no need to create a copy of the URL list in the stack of each recursion. We can keep this data as a list in the class attributes. I don't know if this will have a positive impact in terms of memory consumption. Since we keep one list instead of multiple, there should not be any negative impact.

4) Ensure consistency between runs triggered by the navigation and extras monitors.
When providing all extras at once, NAPPA triggers the Intent Extras monitor a few milliseconds before the navigation monitor. When providing each extra individually, the same can occur, triggering the prefetcher by each extra. With updates (2) and (3), this can cause different runs to conflict with each other. To ensure consistency between runs and keep the data of each run isolated, the visited nodes and selected URLs lists were encapsulated under a map, using an incremental run number as keys. This key is passed down the recursion and each recursion access only its entry in the map. After completing the recursion, the URL list is retrieved and the entries of both lists in the maps are removed.

5) Fixes the score calculation
There was a typo when calculating the scores, resulting in the Greedy score being larger than it should.