Improvements and fixes to TFPR strategy
1) Lower the default threshold to 30%
Preliminary runs with 7 real-world open-source Android apps showed that the majority of the apps calculated a score between 30% to 40% for the navigated activities.

2) The list of best nodes was being sorted from the lowest to the highest
When the SDK was downgraded to level 23, the comparator was incorrectly replaced. This commit fixes the comparator to sort the nodes from the highest to the lowest score.

3) Only run algorithm and scores calculations if needed
If the current node doesn't have any successor, there is no need to generate a subgraph, calculate the aggregate times and run the TFPR algorithm. Prior to executing any of these tasks, verify if the current node has any successor.