Ensures the execution is not interrupted by an Exception
As added in the comments, this is a quick fix. I did not have the time to investigate why a NullPointException was being thrown here in the first place. I assume that we don't want null URL pieces since they don't add any value to NAPPA's prefetching abilities. In future, the initialization and management of URL pieces need to be debugged to verify where null values were accepted and saved.