Remove the TODO workflow.
I add this workflow to automatically parse the pushed source code and create issues from TODO comments. However, after using it during the project, I think that this is not a good solution: creates issues are not well-formatted and it adds clutter to the source code. A better solution is simply using GitHub linking feature and manually creating the issues. This way the issue is linked to the source code as intended, but without inserting TODO notes.